+++ src/TorchSharp/NN/Sequential.cs
@@ -258,6 +258,7 @@ namespace TorchSharp
             /// </remarks>
             public override void train(bool on = true)
             {
+                base.train(on);
                 foreach (var m in _modules) { ((torch.nn.Module)m).train(on); }
             }
 
@@ -269,6 +270,7 @@ namespace TorchSharp
             /// </remarks>
             public override void eval()
             {
+                base.eval();
                 foreach (var m in _modules) { ((torch.nn.Module)m).eval(); }
             }