+++ src/TorchSharp/Torch.cs
@@ -21,13 +21,13 @@ namespace TorchSharp
     {
 #if LIBTORCH_2_2_2_0
         const string libtorchPackageVersion = "2.2.2.0";
-#elif LIBTORCH_2_5_1_0
-        const string libtorchPackageVersion = "2.5.1.0";
+#elif LIBTORCH_2_7_1_0
+        const string libtorchPackageVersion = "2.7.1.0";
 #else
 #error "Please update libtorchPackageVersion to match LibTorchPackageVersion"
 #endif
-#if CUDA_12_1
-        const string cudaVersion = "12.1";
+#if CUDA_12_8
+        const string cudaVersion = "12.8";
 #else
 #error "Please update cudaVersion to match CudaVersionDot"
 #endif
@@ -55,6 +55,26 @@ namespace TorchSharp
         static bool nativeBackendCudaLoaded = false;
 
         public static string __version__ => libtorchPackageVersion;
+        public static string NormalizeNuGetVersion(string versionString)
+        {
+            if (string.IsNullOrWhiteSpace(versionString))
+                throw new ArgumentException($"Invalid NuGet version: {versionString}. Version string is null, empty or only contains whitespaces");
+
+            string[] parts = versionString.Split('-', '+');
+            string[] versionParts = parts[0].Split('.');
+
+            if (versionParts.Length < 2 || versionParts.Length > 4 || !versionParts.All(v => int.TryParse(v, out _)))
+                throw new ArgumentException($"Invalid NuGet version: {versionString}. Please check: https://learn.microsoft.com/en-us/nuget/concepts/package-versioning");
+
+            string normalizedVersion = versionParts[0] + "." + versionParts[1];
+            if (versionParts.Length > 2) normalizedVersion += "." + versionParts[2];
+            if (versionParts.Length > 3 && int.Parse(versionParts[3]) != 0) normalizedVersion += "." + versionParts[3];
+
+            if (parts.Length > 1)
+                normalizedVersion += "-" + parts[1];
+
+            return normalizedVersion;
+        }
 
         internal static bool TryLoadNativeLibraryFromFile(string path, StringBuilder trace)
         {
@@ -121,9 +141,14 @@ namespace TorchSharp
                         ok = TryLoadNativeLibraryByName("cudnn_heuristic64_9.dll", typeof(torch).Assembly, trace);
                         ok = TryLoadNativeLibraryByName("cudnn_engines_precompiled64_9.dll", typeof(torch).Assembly, trace);
                         ok = TryLoadNativeLibraryByName("cudnn_engines_runtime_compiled64_9.dll", typeof(torch).Assembly, trace);
-                        ok = TryLoadNativeLibraryByName("nvrtc-builtins64_121", typeof(torch).Assembly, trace);
+                        ok = TryLoadNativeLibraryByName("nvrtc-builtins64_128", typeof(torch).Assembly, trace);
                         ok = TryLoadNativeLibraryByName("caffe2_nvrtc", typeof(torch).Assembly, trace);
                         ok = TryLoadNativeLibraryByName("nvrtc64_120_0", typeof(torch).Assembly, trace);
+                        ok = TryLoadNativeLibraryByName("cublasLt64_12", typeof(torch).Assembly, trace);
+                        ok = TryLoadNativeLibraryByName("cufft64_11", typeof(torch).Assembly, trace);
+                        ok = TryLoadNativeLibraryByName("fbgemm", typeof(torch).Assembly, trace);
+                        ok = TryLoadNativeLibraryByName("cusparse64_12", typeof(torch).Assembly, trace);
+                        ok = TryLoadNativeLibraryByName("cusolver64_11", typeof(torch).Assembly, trace);
                     }
 
                     ok = TryLoadNativeLibraryByName("torch_cuda", typeof(torch).Assembly, trace);
@@ -170,14 +195,14 @@ namespace TorchSharp
 
                     if (torchsharpLoc!.Contains("torchsharp") && torchsharpLoc.Contains("lib") && Directory.Exists(packagesDir) && Directory.Exists(torchsharpHome)) {
 
-                        var torchSharpVersion = Path.GetFileName(torchsharpHome); // really GetDirectoryName
-
+                        var torchSharpVersion = NormalizeNuGetVersion(Path.GetFileName(torchsharpHome));
+                        var normalizedLibtorchPackageVersion = NormalizeNuGetVersion(libtorchPackageVersion);
                         if (useCudaBackend) {
                             var consolidatedDir = Path.Combine(torchsharpLoc, $"cuda-{cudaVersion}");
 
                             trace.AppendLine($"    Trying dynamic load for .NET/F# Interactive by consolidating native {cudaRootPackage}-* binaries to {consolidatedDir}...");
 
-                            var cudaOk = CopyNativeComponentsIntoSingleDirectory(packagesDir, $"{cudaRootPackage}-*", libtorchPackageVersion, consolidatedDir, trace);
+                            var cudaOk = CopyNativeComponentsIntoSingleDirectory(packagesDir, $"{cudaRootPackage}-*", normalizedLibtorchPackageVersion, consolidatedDir, trace);
                             if (cudaOk) {
                                 cudaOk = CopyNativeComponentsIntoSingleDirectory(packagesDir, "torchsharp", torchSharpVersion, consolidatedDir, trace);
                                 if (cudaOk) {
@@ -195,7 +220,7 @@ namespace TorchSharp
 
                             trace.AppendLine($"    Trying dynamic load for .NET/F# Interactive by consolidating native {cpuRootPackage}-* binaries to {consolidatedDir}...");
 
-                            var cpuOk = CopyNativeComponentsIntoSingleDirectory(packagesDir, cpuRootPackage, libtorchPackageVersion, consolidatedDir, trace);
+                            var cpuOk = CopyNativeComponentsIntoSingleDirectory(packagesDir, cpuRootPackage, normalizedLibtorchPackageVersion, consolidatedDir, trace);
                             if (cpuOk) {
                                 cpuOk = CopyNativeComponentsIntoSingleDirectory(packagesDir, "torchsharp", torchSharpVersion, consolidatedDir, trace);
                                 if (cpuOk) {